internal/aggregate: extract some functions from cmd/tool/slowest
Move the aggregation of slowest tests into an internal package in
anticipation of using these functions for test metrics.

Also fix the 'gotestsum tool slowest' example in the help text.
