+++ cmd/tool/slowest/slowest.go
@@ -5,10 +5,10 @@ import (
 	"io"
 	"io/ioutil"
 	"os"
-	"sort"
 	"time"
 
 	"github.com/spf13/pflag"
+	"gotest.tools/gotestsum/internal/aggregate"
 	"gotest.tools/gotestsum/log"
 	"gotest.tools/gotestsum/testjson"
 )
@@ -73,7 +73,7 @@ predefined statement, --skip-stmt=testing.Short, which uses this Go statement:
 Alternatively, a custom --skip-stmt may be provided as a string:
 
     skip_stmt='
-        if os.Getenv("TEST_FAST") {
+        if os.GetEnv("TEST_FAST") != "" {
             t.Skip("too slow for TEST_FAST")
         }
     '
@@ -118,7 +118,7 @@ func run(opts *options) error {
 		return fmt.Errorf("failed to scan testjson: %v", err)
 	}
 
-	tcs := slowTestCases(exec, opts.threshold)
+	tcs := aggregate.Slowest(exec, opts.threshold)
 	if opts.skipStatement != "" {
 		skipStmt, err := parseSkipStatement(opts.skipStatement)
 		if err != nil {
@@ -131,69 +131,6 @@ func run(opts *options) error {
 	}
 
 	return nil
-}
-
-// slowTestCases returns a slice of all tests with an elapsed time greater than
-// threshold. The slice is sorted by Elapsed time in descending order (slowest
-// test first).
-//
-// If there are multiple runs of a TestCase, all of them will be represented
-// by a single TestCase with the median elapsed time in the returned slice.
-func slowTestCases(exec *testjson.Execution, threshold time.Duration) []testjson.TestCase {
-	if threshold == 0 {
-		return nil
-	}
-	pkgs := exec.Packages()
-	tests := make([]testjson.TestCase, 0, len(pkgs))
-	for _, pkg := range pkgs {
-		pkgTests := aggregateTestCases(exec.Package(pkg).TestCases())
-		tests = append(tests, pkgTests...)
-	}
-	sort.Slice(tests, func(i, j int) bool {
-		return tests[i].Elapsed > tests[j].Elapsed
-	})
-	end := sort.Search(len(tests), func(i int) bool {
-		return tests[i].Elapsed < threshold
-	})
-	return tests[:end]
-}
-
-// collectTestCases maps all test cases by name, and if there is more than one
-// instance of a TestCase, finds the median elapsed time for all the runs.
-//
-// All cases are assumed to be part of the same package.
-func aggregateTestCases(cases []testjson.TestCase) []testjson.TestCase {
-	if len(cases) < 2 {
-		return cases
-	}
-	pkg := cases[0].Package
-	// nolint: prealloc // size is not predictable
-	m := make(map[testjson.TestName][]time.Duration)
-	for _, tc := range cases {
-		m[tc.Test] = append(m[tc.Test], tc.Elapsed)
-	}
-	result := make([]testjson.TestCase, 0, len(m))
-	for name, timing := range m {
-		result = append(result, testjson.TestCase{
-			Package: pkg,
-			Test:    name,
-			Elapsed: median(timing),
-		})
-	}
-	return result
-}
-
-func median(times []time.Duration) time.Duration {
-	switch len(times) {
-	case 0:
-		return 0
-	case 1:
-		return times[0]
-	}
-	sort.Slice(times, func(i, j int) bool {
-		return times[i] < times[j]
-	})
-	return times[len(times)/2]
 }
 
 func jsonfileReader(v string) (io.ReadCloser, error) {
