+++ cmd/tool/slowest/testdata/cmd-flags-help-text
@@ -26,7 +26,7 @@ predefined statement, --skip-stmt=testing.Short, which uses this Go statement:
 Alternatively, a custom --skip-stmt may be provided as a string:
 
     skip_stmt='
-        if os.Getenv("TEST_FAST") {
+        if os.GetEnv("TEST_FAST") != "" {
             t.Skip("too slow for TEST_FAST")
         }
     '
